+++ app/components/design_system/govuk/action_link_component.rb
@@ -1,0 +1,8 @@
+module DesignSystem
+  module Govuk
+    # GOV.UK doesn't have a dedicated action-link pattern, so we render a
+    # button-styled link.
+    class ActionLinkComponent < DesignSystem::Generic::ActionLinkComponent
+    end
+  end
+end