+++ app/components/design_system/govuk/callout_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class CalloutComponent < DesignSystem::Generic::CalloutComponent
+    end
+  end
+end