+++ app/components/design_system/govuk/details_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class DetailsComponent < DesignSystem::Generic::DetailsComponent
+    end
+  end
+end