+++ app/components/design_system/govuk/heading_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class HeadingComponent < DesignSystem::Generic::HeadingComponent
+    end
+  end
+end