+++ app/components/design_system/govuk/list_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class ListComponent < DesignSystem::Generic::ListComponent
+    end
+  end
+end