+++ app/components/design_system/govuk/panel_component.rb
@@ -1,0 +1,10 @@
+module DesignSystem
+  module Govuk
+    # GOV.UK confirmation panel.
+    class PanelComponent < DesignSystem::Generic::PanelComponent
+      def panel_classes
+        super + ["#{brand}-panel--confirmation"]
+      end
+    end
+  end
+end