+++ app/components/design_system/govuk/start_button_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class StartButtonComponent < DesignSystem::Generic::StartButtonComponent
+    end
+  end
+end