+++ app/components/design_system/nhsuk/heading_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Nhsuk
+    class HeadingComponent < DesignSystem::Generic::HeadingComponent
+    end
+  end
+end