+++ lib/design_system/generic.rb
@@ -2,20 +2,14 @@
 
 # This is the generic (unbranded) adapter for design system
 
-require_relative 'generic/builders/concerns/brand_derivable'
 require_relative 'generic/builders/button'
-require_relative 'generic/builders/callout'
 require_relative 'generic/builders/code'
-require_relative 'generic/builders/details'
 require_relative 'generic/builders/fixed_elements'
 require_relative 'generic/builders/grid'
-require_relative 'generic/builders/heading'
 require_relative 'generic/builders/inset_text'
 require_relative 'generic/builders/link'
-require_relative 'generic/builders/list'
 require_relative 'generic/builders/notification'
 require_relative 'generic/builders/pagination_renderer'
-require_relative 'generic/builders/panel'
 require_relative 'generic/builders/paragraph'
 require_relative 'generic/builders/tab'
 require_relative 'generic/builders/table'